SAK-34049 Initial rubrics integration
