+++ rubrics/api/src/java/org/sakaiproject/rubrics/logic/api/model/BaseResource.java
@@ -20,7 +20,7 @@
  *
  **********************************************************************************/
 
-package org.sakaiproject.rubrics.model;
+package org.sakaiproject.rubrics.logic.api.model;
 
 
 import lombok.Data;
