+++ rubrics/api/src/java/org/sakaiproject/rubrics/logic/api/model/Rating.java
@@ -20,7 +20,7 @@
  *
  **********************************************************************************/
 
-package org.sakaiproject.rubrics.model;
+package org.sakaiproject.rubrics.logic.api.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
@@ -28,49 +28,39 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinTable;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.OrderColumn;
 import javax.persistence.PostLoad;
 import javax.persistence.PostUpdate;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
 import java.io.Serializable;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "rbc_criterion")
-@JsonPropertyOrder({"id", "title", "description", "metadata"})
-public class Criterion extends BaseResource<Criterion.Metadata> implements Serializable, Cloneable {
+@Table(name = "rbc_rating")
+@JsonPropertyOrder({"id", "title", "description", "points", "metadata"})
+public class Rating extends BaseResource<Rating.Metadata> implements Serializable, Cloneable {
 
     @Id
-    @SequenceGenerator(name="rbc_crit_seq", sequenceName="rbc_crit_seq")
-    @GeneratedValue(strategy = GenerationType.AUTO, generator = "rbc_crit_seq")
+    @SequenceGenerator(name="rbc_rat_seq", sequenceName ="rbc_rat_seq")
+    @GeneratedValue(strategy = GenerationType.AUTO, generator = "rbc_rat_seq")
     private Long id;
 
     private String title;
     private String description;
-
-    @OneToMany(cascade = CascadeType.ALL)
-    @JoinTable(name = "rbc_criterion_ratings")
-    @OrderColumn(name = "order_index")
-    private List<Rating> ratings;
+    private Integer points;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JsonIgnore
-    private Rubric rubric;
+    private Criterion criterion;
 
     public Metadata getMetadata() {
         if (this.metadata == null) {
@@ -85,27 +75,22 @@ public class Criterion extends BaseResource<Criterion.Metadata> implements Seria
     @PostLoad
     @PostUpdate
     public void determineSharedParentStatus() {
-        Rubric rubric = getRubric();
-        if (rubric != null && rubric.getMetadata().isShared()) {
-            getMetadata().setShared(true);
+        Criterion criterion = getCriterion();
+        if (criterion != null) {
+            Rubric rubric = criterion.getRubric();
+            if (rubric != null && rubric.getMetadata().isShared()) {
+                getMetadata().setShared(true);
+            }
         }
     }
 
     @Override
-    public Criterion clone() throws CloneNotSupportedException {
-        Criterion clonedCriterion = new Criterion();
-        clonedCriterion.setId(null);
-        clonedCriterion.setTitle(this.title);
-        clonedCriterion.setDescription(this.description);
-        clonedCriterion.setRatings(this.getRatings().stream().map(rating -> {
-            Rating clonedRating = null;
-            try {
-                clonedRating = rating.clone();
-            } catch (CloneNotSupportedException e) {
-                throw new RuntimeException(e);
-            }
-            return clonedRating;
-        }).collect(Collectors.toList()));
-        return clonedCriterion;
+    public Rating clone() throws CloneNotSupportedException {
+        Rating clonedRating = new Rating();
+        clonedRating.setId(null);
+        clonedRating.setTitle(this.title);
+        clonedRating.setDescription(this.description);
+        clonedRating.setPoints(this.points);
+        return clonedRating;
     }
 }
