+++ rubrics/api/src/java/org/sakaiproject/rubrics/logic/api/model/projections/InlineRubric.java
@@ -20,19 +20,20 @@
  *
  **********************************************************************************/
 
-package org.sakaiproject.rubrics.model.projections;
+package org.sakaiproject.rubrics.logic.api.model.projections;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import org.sakaiproject.rubrics.model.Criterion;
-import org.sakaiproject.rubrics.model.Rating;
-import org.sakaiproject.rubrics.model.Rubric;
+
+import org.sakaiproject.rubrics.logic.api.model.Rubric;
+import org.sakaiproject.rubrics.logic.api.model.ToolItemRubricAssociation;
 import org.springframework.data.rest.core.config.Projection;
 
 import java.util.List;
 
-@Projection(name = "inlineCriterion", types = { Criterion.class })
-@JsonPropertyOrder({"id", "title", "description", "metadata", "ratings"})
-public interface InlineCriterion {
+@Projection(name = "inlineRubric", types = { Rubric.class })
+@JsonPropertyOrder({"id", "title", "description", "metadata", "criterions"})
+public interface InlineRubric {
 
     Long getId();
 
@@ -40,7 +41,10 @@ public interface InlineCriterion {
 
     String getDescription();
 
-    List<Rating> getRatings();
+    List<InlineCriterion> getCriterions();
 
-    Criterion.Metadata getMetadata();
+    Rubric.Metadata getMetadata();
+
+    @JsonIgnore
+    List<ToolItemRubricAssociation> getToolItemAssociations();
 }
