--- rubrics/api/src/java/org/sakaiproject/rubrics/model/ToolItemRubricAssociation.java
@@ -1,61 +1,0 @@
-/**********************************************************************************
- *
- * Copyright (c) 2017 The Sakai Foundation
- *
- * Original developers:
- *
- *   Unicon
- *
- * Licensed under the Educational Community License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.osedu.org/licenses/ECL-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- **********************************************************************************/
-
-package org.sakaiproject.rubrics.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-import java.time.Instant;
-import java.util.Map;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ToolItemRubricAssociation implements Serializable {
-
-    private Long id;
-    private String toolId;
-    private String itemId;
-    private Long rubricId;
-    private Rubric rubric;
-    private Metadata metadata;
-
-    @Data
-    public static class Metadata {
-        private String created;
-        private String modified;
-        private String owner;
-    }
-    private Map<String, Boolean> parameters;
-
-    public Map<String, Boolean> getParameters() {
-            return parameters;
-        }
-
-    public Boolean getParameter(String param) {
-            return parameters.get(param);
-        }
-
-}
