+++ rubrics/impl/src/webapp/WEB-INF/components.xml
@@ -17,23 +17,5 @@
         <property name="siteService" ref="org.sakaiproject.site.api.SiteService" />
         <property name="functionManager" ref="org.sakaiproject.authz.api.FunctionManager" />
         <property name="authzGroupService" ref="org.sakaiproject.authz.api.AuthzGroupService"/>
-        <property name="cache" ref="org.sakaiproject.rubrics.cache" />
     </bean>
-
-    <!-- Cache: ttl=1 hour -->
-    <bean id="org.sakaiproject.rubrics.cache"
-        class="org.springframework.cache.ehcache.EhCacheFactoryBean">
-        <property name="cacheManager">
-            <ref bean="org.sakaiproject.memory.api.MemoryService.cacheManager"/>
-        </property>
-        <property name="cacheName">
-            <value>org.sakaiproject.rubrics.cache</value>
-        </property>
-        <property name="maxElementsInMemory" value="1000" />
-        <property name="diskPersistent" value="false" />
-        <property name="eternal" value="false" />
-        <property name="timeToLive" value="3600" />
-        <property name="timeToIdle" value="3600" />
-    </bean>
-
 </beans>
