+++ rubrics/services/src/main/resources/application.properties
@@ -1,9 +1,9 @@
 # See http://docs.spring.io/spring-boot/docs/current/reference/html/common-application-properties.html
 spring.thymeleaf.cache=false
 spring.main.show-banner=false
-logging.level.jdbc=OFF
-logging.level.jdbc.sqltiming=DEBUG
-logging.level.jdbc.resultsettable=DEBUG
+#logging.level.jdbc=OFF
+#logging.level.jdbc.sqltiming=DEBUG
+#logging.level.jdbc.resultsettable=DEBUG
 
 server.servlet-path=/rest
 #spring.data.rest.base-path=rest
@@ -12,20 +12,21 @@ spring.data.rest.returnBodyOnCreate=true
 spring.data.rest.returnBodyOnUpdate=true
 
 spring.datasource.initialize=true
-
-spring.datasource.url=jdbc:mysql://localhost/sakai_usu
-spring.datasource.username=sakai_usu
-spring.datasource.password=sakai_usu
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.datasource.tomcat.test-on-borrow=true
-spring.datasource.tomcat.validation-query=SELECT 1
-spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
-spring.jpa.generate-ddl=true
+#spring.datasource.type=javax.sql.LazyDataSource
+#spring.datasource.tomcat.data-source=javax.sql.LazyDataSource
+#spring.datasource.url=jdbc:mysql://localhost/sakai_usu
+#spring.datasource.username=sakai_usu
+#spring.datasource.password=sakai_usu
+#spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+#spring.datasource.tomcat.test-on-borrow=true
+#spring.datasource.tomcat.validation-query=SELECT 1
+#spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
+#spring.jpa.generate-ddl=true
 # Automatically validates or exports schema DDL to the database when the SessionFactory is created.
 # With create-drop, the database schema will be dropped when the SessionFactory is closed explicitly.
 # Possible values: validate | update | create | create-drop
 #spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.hibernate.ddl-auto=create-drop
+#spring.jpa.hibernate.ddl-auto=update
 
 # ###############################
 # Rubrics
