--- rubrics/services/src/main/resources/log4jdbc.log4j2.properties
@@ -1,2 +1,0 @@
-log4jdbc.dump.sql.maxlinelength=0
-log4jdbc.spylogdelegator.name=net.sf.log4jdbc.log.slf4j.Slf4jSpyLogDelegator
